Numeric Mutator Base Class

Base class for mutation operations on numeric and integer valued individuals, inheriting from Mutator.

MutatorNumeric operators perform mutation on numeric (integer and real valued) individuals. Inheriting operators implement the private $.mutate_numeric() function that is called once for each individual and is given a numeric vector.

Inheriting

MutatorNumeric is an abstract base class and should be inherited from. Inheriting classes should implement the private $.mutate_numeric() function. During $operate(), the $.mutate_numeric() function is called once for each individual, with the parameters values (the individual as a single numeric vector), lowers and uppers (numeric vectors, the lower and upper bounds for each component of values). Typically, $initialize() should also be overloaded.

MutatorNumerics that perform real-valued operations, such as e.g. MutatorGauss, operate on integers by widening the lower and upper bounds of integer components by 0.5, applying their operation, and rounding resulting values to the nearest integer (while always staying inside bounds).

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Mutator -> MutatorNumeric

Methods

Public methods

MutatorNumeric$new()

MutatorNumeric$clone()

Inherited methods

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$prime()

Method new()

Initialize base class components of the MutatorNumeric.

Usage

MutatorNumeric$new(param_classes = c("ParamInt", "ParamDbl"), param_set = ps())

Arguments

param_classes

(character)
Classes of parameters that the operator can handle. May contain any of "ParamInt", "ParamDbl". Default is both of them.
The $param_classes field will reflect this value.

param_set

(ParamSet | list of expression)
Strategy parameters of the operator. This should be created by the subclass and given to super$initialize(). If this is a ParamSet, it is used as the MiesOperator's ParamSet directly. Otherwise it must be a list of expressions e.g. created by alist() that evaluate to ParamSets, possibly referencing self and private. These ParamSet are then combined using a ParamSetCollection. Default is the empty ParamSet.
The $param_set field will reflect this value.

Method clone()

The objects of this class are cloneable with this method.

Usage

MutatorNumeric$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.