Mixed Integer Evolutionary Strategies Optimizer

Perform optimization using evolutionary strategies. OptimizerMies and TunerMies implement a standard ES optimization algorithm, performing initialization first, followed by a loop of performance evaluation, survival selection, parent selection, mutation, and recombination. Currently two different survival modes ("comma" and "plus") are supported. Multi-fidelity optimization, similar to the "rolling-tide" algorithm described in Fieldsend (2014). The modular design and reliance on MiesOperator objects to perform central parts of the optimization algorithm makes this Optimizer highly flexible and configurable. In combination with OperatorCombination mutators and recombinators, an algorithm as presented in Li (2013) can easily be implemented.

OptimizerMies implements a standard evolutionary strategies loop:

Prime operators, using mies_prime_operators()

Initialize and evaluate population, using mies_init_population()

Optionally, evaluate survivors with higher fidelity if the multi-fidelity functionality is being used

Generate offspring by selecting parents, recombining and mutating them, using mies_generate_offspring()

Evaluate performance, using mies_evaluate_offspring()

Select survivors, using either mies_survival_plus() or mies_survival_comma(), depending on the survival_strategy configuration parameter

Jump to 3.

Terminating

As with all optimizers, Terminators are used to end optimization after a specific number of evaluations were performed, time elapsed, or other conditions are satisfied. Of particular interest is TerminatorGenerations, which terminates after a number of generations were evaluated in OptimizerMies. The initial population counts as generation 1, its offspring as generation 2 etc.; fidelity refinements are always included in their generation, TerminatorGenerations avoids terminating right before they are evaluated. Other terminators may, however, end the optimization process at any time.

Multi-Fidelity

miesmuschel provides a simple multi-fidelity optimization mechanism that allows increasing fidelity both by generation number and survival status. When multi_fidelity is TRUE, then one search space component of the OptimInstance must have the "budget" tag, which is then optimized as the "budget" component. This means that the value of this component is determined by the fidelity_schedule configuration parameter, which must contain a data.frame with columns "generation", "budget_new" and "budget_survivors". The budget component's value of newly sampled individuals is set to the "budget_new" entry in the generation's row, and surviving individuals are evaluated again with the budget set to the "budget_survivors" value (unless they are the same). At the end of a generation, if the "budget_survivors" value changes, all individuals from previous generations are re-evaluated with the new budget value, unless fidelity_current_gen_only is set to TRUE (and unless the budget value decreases and fidelity_monotonic is TRUE). This makes it possible to implement increasing fidelity by generation, and also increasing fidelity for samples that survived a generation.

The fidelity_schedule configuration parameter's "generation" column determines which row is currently active. A row becomes active in the generation that is listed, and becomes inactive whenever a different row becomes active. So e.g. if fidelity_schedule contains a row with "generation" set to 1, and one set to 4, then the first row is active during generations 1, 2, and 3, and the second row is active for all following generations.

Note that the multifidelity functionality is experimental and the UI may change in the future.

Additional Components

The search space over which the optimization is performed is fundamentally tied to the Objective, and therefore to the OptimInstance given to OptimizerMies$optimize(). However, some advanced Evolutionary Strategy based algorithms may need to make use of additional search space components that are independent of the particular objective. An example is self-adaption as implemented in OperatorCombination, where one or several components can be used to adjust operator behaviour. These additional components are supplied to the optimizer through the additional_component_sampler configuration parameter, which takes a Sampler object. This object both has an associated ParamSet which represents the additional components that are present, and it provides a method for generating the initial values of these components. The search space that is seen by the MiesOperators is then the union of the OptimInstance's ParamSet, and the Sampler's ParamSet.

Configuration Parameters

OptimizerMies has the configuration parameters of the mutator, recombinator, parent_selector, survival_selector, and, if given, elite_selector operator given during construction, and prefixed according to the name of the argument (mutator's configuration parameters are prefixed "mutator." etc.). When using the construction arguments' default values, they are all "proxy" operators: MutatorProxy, RecombinatorProxy and SelectorProxy. This means that the respective configuration parameters become mutator.operation, recombinator.operation etc., so the operators themselves can be set via configuration parameters in this case.

Further configuration parameters are:

lambda :: integer(1)
Offspring size: Number of individuals that are created and evaluated anew for each generation. This is equivalent to the lambda parameter of mies_generate_offspring(), see there for more information. Initialized to 10.

mu :: integer(1)
Population size: Number of individuals that are sampled in the beginning, and which are selected with each survival step. This is equivalent to the mu parameter of mies_init_population(), see there for more information. Initialized to 1.

survival_strategy :: character(1)
May be "plus", or, if the elite_selector construction argument is not NULL, "comma": Choose whether mies_survival_plus() or mies_survival_comma() is used for survival selection. Initialized to "plus".

n_elite :: integer(1)
Only if the elite_selector construction argument is not NULL, and only valid when survival_strategy is "comma": Number of elites, i.e. individuals from the parent generation, to keep during "Comma" survival. This is equivalent to the n_elite parameter of mies_survival_comma(), see there for more information.

initializer :: function
Function that generates the initial population as a Design object, with arguments param_set and n, functioning like paradox::generate_design_random or paradox::generate_design_lhs. This is equivalent to the initializer parameter of mies_init_population(), see there for more information. Initialized to generate_design_random().

additional_component_sampler :: Sampler | NULL
Additional components that may be part of individuals as seen by mutation, recombination, and selection MiesOperators, but that are not part of the search space of the OptimInstance being optimized. This is equivalent to the additional_component_sampler parameter of mies_init_population(), see there for more information. Initialized to NULL (no additional components).

fidelity_schedule :: data.frame
Only if the multi_fidelity construction argument is TRUE: Table that determines the value of the "budget" component of individuals being evaluated when doing multi-fidelity optimization. This is equivalent to the fidelity_schedule parameter of mies_init_population(), mies_evaluate_offspring(), and mies_step_fidelity(); see there for more information.
When this configuration parameter is present (i.e. multi_fidelity is TRUE), then it is initialized to a data.frame containing one row for generation 1, setting budget to 1 for both new and survivor individuals.
Note that the multifidelity functionality is experimental and the UI may change in the future.

fidelity_generation_lookahead :: logical(1)
Only if the multi_fidelity construction argument is TRUE: Whether to use the "survivor_budget" of the next generation, instead of the current generation, when doing fidelity refinement in mies_step_fidelity(). This is equivalent to the generation_lookahead parameter of mies_step_fidelity(), see there for more information.
When this configuration parameter is present (i.e. multi_fidelity is TRUE), then it is initialized to TRUE.
Note that the multifidelity functionality is experimental and the UI may change in the future.

fidelity_current_gen_only :: logical(1)
Only if the multi_fidelity construction argument is TRUE: When doing fidelity refinement in mies_step_fidelity(), whether to refine all individuals with different budget component, or only individuals created in the current generation. This is equivalent to the current_gen_only parameter of mies_step_fidelity(), see there for more information.
When this configuration parameter is present (i.e. multi_fidelity is TRUE), then it is initialized to FALSE.
Note that the multifidelity functionality is experimental and the UI may change in the future.

fidelity_monotonic :: logical(1)
Only if the multi_fidelity construction argument is TRUE: Whether to only do fidelity refinement in mies_step_fidelity() for individuals for which the when budget component value would increase. This is equivalent to the monotonic parameter of mies_step_fidelity(), see there for more information.
When this configuration parameter is present (i.e. multi_fidelity is TRUE), then it is initialized to TRUE.
Note that the multifidelity functionality is experimental and the UI may change in the future.

References

Fieldsend, E J, Everson, M R (2014). “The rolling tide evolutionary algorithm: A multiobjective optimizer for noisy optimization problems.” IEEE Transactions on Evolutionary Computation, 19(1), 103--117.

Li, Rui, Emmerich, TM M, Eggermont, Jeroen, B"ack, Thomas, Sch"utz, Martin, Dijkstra, Jouke, Reiber, HC J (2013). “Mixed integer evolution strategies for parameter optimization.” Evolutionary computation, 21(1), 29--64.

See also

Super class

bbotk::Optimizer -> OptimizerMies

Active bindings

mutator

(Mutator)
Mutation operation to perform during mies_generate_offspring().

recombinator

(Recombinator)
Recombination operation to perform during mies_generate_offspring().

parent_selector

(Selector)
Parent selection operation to perform during mies_generate_offspring().

survival_selector

(Selector)
Survival selection operation to use in mies_survival_plus() or mies_survival_comma().

elite_selector

(Selector | NULL)
Elite selector used in mies_survival_comma().

param_set

(ParamSet)
Configuration parameters of the optimization algorithm.

Methods

Public methods

OptimizerMies$new()

OptimizerMies$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Initialize the OptimizerMies object.

Usage

OptimizerMies$new(
  mutator = MutatorProxy$new(),
  recombinator = RecombinatorProxy$new(),
  parent_selector = SelectorProxy$new(),
  survival_selector = SelectorProxy$new(),
  elite_selector = NULL,
  multi_fidelity = FALSE
)

Arguments

mutator

(Mutator)
Mutation operation to perform during mies_generate_offspring(), see there for more information. Default is MutatorProxy, which exposes the operation as a configuration parameter of the optimizer itself.
The $mutator field will reflect this value.

recombinator

(Recombinator)
Recombination operation to perform during mies_generate_offspring(), see there for more information. Default is RecombinatorProxy, which exposes the operation as a configuration parameter of the optimizer itself. Note: The default RecombinatorProxy has $n_indivs_in set to 2, so to use recombination operations with more than two inputs, or to use population size of 1, it may be necessary to construct this argument explicitly.
The $recombinator field will reflect this value.

parent_selector

(Selector)
Parent selection operation to perform during mies_generate_offspring(), see there for more information. Default is SelectorProxy, which exposes the operation as a configuration parameter of the optimizer itself.
The $parent_selector field will reflect this value.

survival_selector

(Selector)
Survival selection operation to use in mies_survival_plus() or mies_survival_comma() (depending on the survival_strategy configuration parameter), see there for more information. Default is SelectorProxy, which exposes the operation as a configuration parameter of the optimizer itself.
The $survival_selector field will reflect this value.

elite_selector

(Selector | NULL)
Elite selector used in mies_survival_comma(), see there for more information. "Comma" selection is only available when this argument is not NULL. Default NULL.
The $elite_selector field will reflect this value.

multi_fidelity

(logical(1))
Whether to enable multi-fidelity optimization. When this is TRUE, then the OptimInstance being optimized must contain a Param tagged "budget", which is then used as the "budget" search space component, determined by fidelity_schedule instead of by the MiesOperators themselves. For multi-fidelity optimization, the fidelity_schedule, fidelity_generation_lookahead, fidelity_current_gen_only, and fidelity_monotonic configuration parameters must be given to determine multi-fidelity behaviour. (While the initial values for most of these are probably good for most cases in which more budget implies higher fidelity, the fidelity_schedule configuration parameter should be adjusted in most cases). Default is FALSE.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerMies$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerMies

Methods

Public methods

TunerMies$new()

TunerMies$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Initialize the TunerMies object.

Usage

TunerMies$new(
  mutator = MutatorProxy$new(),
  recombinator = RecombinatorProxy$new(),
  parent_selector = SelectorProxy$new(),
  survival_selector = SelectorProxy$new(),
  elite_selector = NULL,
  multi_fidelity = FALSE
)

Arguments

mutator

(Mutator)

recombinator

(Recombinator)

parent_selector

(Selector)

survival_selector

(Selector)

elite_selector

(Selector | NULL)

multi_fidelity

(logical(1))

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerMies$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
lgr::threshold("warn")

op.m <- mut("gauss")
op.r <- rec("xounif", p = .3)
op.parent <- sel("random")
op.survival <- sel("best")

#####
# Optimizing a Function
#####

library("bbotk")

# Define the objective to optimize
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

# Get a new OptimInstance
oi <- OptimInstanceSingleCrit$new(objective,
  terminator = trm("evals", n_evals = 100)
)

# Create OptimizerMies object
mies_opt <- opt("mies", mutator = op.m, recombinator = op.r,
  parent_selector = op.parent, survival_selector = op.survival,
  mu = 10, lambda = 5)

# mies_opt$optimize performs MIES optimization and returns the optimum
mies_opt$optimize(oi)

#>           x        y  x_domain      Obj
#> 1: 2.280652 1.801712 <list[2]> 1.777458

#####
# Optimizing a Machine Learning Method
#####

# Note that this is a short example, aiming at clarity and short runtime.
# The settings are not optimal for hyperparameter tuning. The resampling
# in particular should not be "holdout" for small datasets where this gives
# a very noisy estimate of performance.

library("mlr3")
library("mlr3tuning")

# The Learner to optimize
learner = lrn("classif.rpart")

# The hyperparameters to optimize
learner$param_set$values[c("cp", "maxdepth")] = list(to_tune())

# Get a TuningInstance
ti = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.acc"),
  terminator = trm("gens", generations = 10)
)

# Create TunerMies object
mies_tune <- tnr("mies", mutator = op.m, recombinator = op.r,
  parent_selector = op.parent, survival_selector = op.survival,
  mu = 10, lambda = 5)

# mies_tune$optimize performs MIES optimization and returns the optimum
mies_tune$optimize(ti)

#>            cp maxdepth learner_param_vals  x_domain classif.acc
#> 1: 0.08492106        6          <list[3]> <list[2]>         0.9

# }