Recombinator Base Class

Base class representing recombination operations, inheriting from MiesOperator.

Recombinators get a table of individuals as input and return a table of modified individuals as output. Individuals are acted on by groups: every $n_indivs_out lines of output corresponds to a group of $n_indivs_in lines of input, and presence or absence of other input groups does not affect the result.

Recombination operations are performed in ES algorithms to facilitate exploration of the search space that combine partial solutions.

Inheriting

Recombinator is an abstract base class and should be inherited from. Inheriting classes should implement the private $.recombine() function. The user of the object calls $operate(), which calls $.recombine() for each $n_indivs_in sized group of individuals after checking that the operator is primed, that the values argument conforms to the primed domain. $.recombine() should then return a table of $n_indivs_out individuals for each call. Typically, the $initialize() function should also be overloaded, and optionally the $prime() function; they should call their super equivalents.

See also

Super class

miesmuschel::MiesOperator -> Recombinator

Active bindings

n_indivs_in

(integer(1))
Number of individuals to consider at the same time. When operating, the number of input individuals must be divisible by this number.

n_indivs_out

(integer(1))
Number of individuals produced for each group of $n_indivs_in individuals.

Methods

Public methods

Recombinator$new()

Recombinator$clone()

Inherited methods

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$prime()

Method new()

Initialize base class components of the Recombinator.

Usage

Recombinator$new(
  param_classes = c("ParamLgl", "ParamInt", "ParamDbl", "ParamFct"),
  param_set = ps(),
  n_indivs_in = 2,
  n_indivs_out = n_indivs_in
)

Arguments

param_classes

(character)
Classes of parameters that the operator can handle. May contain any of "ParamLgl", "ParamInt", "ParamDbl", "ParamFct". Default is all of them.
The $param_classes field will reflect this value.

param_set

(ParamSet | list of expression)
Strategy parameters of the operator. This should be created by the subclass and given to super$initialize(). If this is a ParamSet, it is used as the MiesOperator's ParamSet directly. Otherwise it must be a list of expressions e.g. created by alist() that evaluate to ParamSets, possibly referencing self and private. These ParamSet are then combined using a ParamSetCollection. Default is the empty ParamSet.
The $param_set field will reflect this value.

n_indivs_in

(integer(1))
Number of individuals to consider at the same time. When operating, the number of input individuals must be divisible by this number. Default 2.
The $n_indivs_in field will reflect this value.

n_indivs_out

(integer(1))
Number of individuals that result for each n_indivs_in lines of input. The number of results from the recombinator will be nrow(values) / n_indivs_in * n_indivs_out. Default equal to n_indivs_in.
The $n_indivs_out field will reflect this value.

Method clone()

The objects of this class are cloneable with this method.

Usage

Recombinator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.