Sampler for Projection Weights

Sampler for a single ParamUty that samples weight-matrices as used by ScalorFixedProjection.

Super class

paradox::Sampler -> SamplerRandomWeights

Active bindings

nobjectives

(numeric(1))
Number of objectives for which weights are generated.

nweights

(numeric(1))
Number of weight vectors generated for each configuration.

weights_component_id

(numeric(1))
search space component identifying the weights by which to scalarize.

Methods

Public methods

SamplerRandomWeights$new()

SamplerRandomWeights$clone()

Inherited methods

paradox::Sampler$format()

paradox::Sampler$print()

paradox::Sampler$sample()

Method new()

Initialize the SamplerRandomWeights object.

Usage

SamplerRandomWeights$new(
  nobjectives = 2,
  nweights = 1,
  weights_component_id = "scalarization_weights"
)

Arguments

nobjectives

(numeric(1))
Number of objectives for which weights are generated.

nweights

(numeric(1))
Number of weight vectors generated for each configuration.

weights_component_id

(character(1))
Id of the ParamUty. Default is "scalarization_weights". Can be changed arbitrarily but should match the ScalorFixedProjection's weights_component_id.

Method clone()

The objects of this class are cloneable with this method.

Usage

SamplerRandomWeights$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

set.seed(1)