Null-Filtor

Null-filtor that does not perform filtering. Its needed_input() is always the output_size, and operate() selects the first n_filter values from its input.

Useful in particular with operator-wrappers such as FiltorProxy, and to make filtering optional.

Configuration Parameters

This operator has no configuration parameters.

Supported Operand Types

Supported Param classes are: ParamLgl, ParamInt, ParamDbl, ParamFct

Dictionary

This Filtor can be created with the short access form ftr() (ftrs() to get a list), or through the the dictionary dict_filtors in the following way:

# preferred:
ftr("null")
ftrs("null")  # takes vector IDs, returns list of Filtors

# long form:
dict_filtors$get("null")

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Filtor -> FiltorNull

Methods

Public methods

FiltorNull$new()

FiltorNull$clone()

Inherited methods

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$prime()

miesmuschel::Filtor$needed_input()

Method new()

Usage

FiltorNull$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FiltorNull$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.