Proxy-Filtor that Filters According to its Configuration Parameter

Filtor that performs the operation in its operation configuration parameter. This can be used to make filtor operations fully parametrizable.

Configuration Parameters

operation :: Filtor
Operation to perform. Must be set by the user. This is primed when $prime() of SelectorProxy is called, and also when $operate() is called, to make changing the operation as part of self-adaption possible. However, if the same operation gets used inside multiple SelectorProxy objects, then it is recommended to $clone(deep = TRUE) the object before assigning them to operation to avoid frequent re-priming.

Supported Operand Types

Supported Param classes are: ParamLgl, ParamInt, ParamDbl, ParamFct

Dictionary

This Selector can be created with the short access form sel() (sels() to get a list), or through the the dictionary dict_selectors in the following way:

# preferred:
sel("proxy")
sels("proxy")  # takes vector IDs, returns list of Selectors

# long form:
dict_selectors$get("proxy")

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Filtor -> FiltorProxy

Methods

Public methods

FiltorProxy$new()

FiltorProxy$prime()

FiltorProxy$clone()

Inherited methods

miesmuschel::MiesOperator$help()

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$print()

miesmuschel::MiesOperator$repr()

miesmuschel::Filtor$needed_input()

Method new()

Initialize the FiltorProxy object.

Usage

FiltorProxy$new()

Method prime()

See MiesOperator method. Primes both this operator, as well as the operator given to the operation configuration parameter. Note that this modifies the $param_set$values$operation object.

Usage

FiltorProxy$prime(param_set)

Arguments

param_set

(ParamSet)
Passed to MiesOperator$prime().

Returns

invisible self.

Method clone()

The objects of this class are cloneable with this method.

Usage

FiltorProxy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("mlr3")
library("mlr3learners")
fp = ftr("proxy")
p = ps(x = p_dbl(-5, 5))
known_data = data.frame(x = 1:5)
fitnesses = 1:5
new_data = data.frame(x = c(2.5, 4.5))

fp$param_set$values$operation = ftr("null")
fp$prime(p)
fp$operate(new_data, known_data, fitnesses, 1)
#> [1] 1

fp$param_set$values$operation = ftr("surprog", lrn("regr.lm"), filter.pool_factor = 2)
fp$operate(new_data, known_data, fitnesses, 1)
#> [1] 2