Progressive Surrogate Model Filtering

Performs progressive surrogate model filtering.

A surrogate model is a regression model, based on an mlr3::Learner, which predicts the approximate performance of newly sampled configurations given the empirical performance of already evaluated configurations. If the optional surrogate_learner construction argument is given to SumoHB, then the surrogate model is used to propose points that have, according to the surrogate model, a relatively high chance of performing well.

The filtering is "progressive" in that successive values are filtered more agressively.

Algorithm

Given the number n_filter of of individuals to sample, progressive surrogate model filtering proceeds as follows:

Train the surrogate_learner LearnerRegr on the known_values and their fitnesses.

Take filter_pool_first configurations, predict their expected performance using the surrogate model, and put them into a pool P of configurations to consider.

Take the individual that is optimal according to predicted performance, remove it from P and add it to solution set S.

If the number of solutions in S equals n_filter, quit.

Take the next filter_pool_per_sample configurations, predict their expected performance using the surrogate model, and add them to P.

Jump to 3.

(The algorithm presented here is optimized for clarity; the actual implementation does all the surrogate model prediction in one go, but is functionally equivalent).

The filter_pool_first and filter_pool_per_sample configuration parameters of this algorithm determine how agressively the surrogate model is used to filter out sampled configurations. If the filtering is agressive (filter_pool_first is large), then more "exploitation" at the cost of "exploration" is performed. When filter_pool_first is small but filter_pool_per_sample is large, then successive individuals are filtered successively more agressively, potentially leading to a tradeoff between "exploration" and "exploitation".

When filter_pool_per_sample is set to 0, then the method is equivalent to sampling the top n_filter individuals from filter_pool_first sampled ones. When filter_pool_per_sample is 1 and filter_pool_first is 0, then the method is equivalent to random sampling.

filter_pool_first and filter_pool_per_sample may be fractional; the total number of individuals to select from when selecting the ith individuals is always round(filter_pool_first + (filter_pool_per_sample - 1) * (i - 1)). However, filter_pool_first must be at least 1, and filter_pool_first + filter_pool_per_sample * (n_filter - 1) must be at least n_filter.

Configuration Parameters

FiltorSurrogateProgressive's configuration parameters are the hyperparameters of the surrogate_learner Learner, as well as:

Supported Operand Types

Supported Param classes depend on the supported feature types of the surrogate_learner, as reported by surrogate_learner$feature_types: "ParamInt" requires "integer", "ParamDbl" requires "numeric", "ParamLgl" requires "logical", and "ParamFct" requires "factor".

Dictionary

This Filtor can be created with the short access form ftr() (ftrs() to get a list), or through the the dictionary dict_filtors in the following way:

# preferred:
ftr("surprog", <surrogate_learner>)
ftrs("surprog", <surrogate_learner>)  # takes vector IDs, returns list of Filtors

# long form:
dict_filtors$get("surprog", <surrogate_learner>)

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Filtor -> FiltorSurrogateProgressive

Active bindings

surrogate_learner

(mlr3::LearnerRegr | NULL)
Regression learner for the surrogate model filtering algorithm.

Methods

Public methods

FiltorSurrogateProgressive$new()

FiltorSurrogateProgressive$clone()

Inherited methods

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$prime()

miesmuschel::Filtor$needed_input()

Method new()

Usage

FiltorSurrogateProgressive$new(surrogate_learner)

Arguments

surrogate_learner

(mlr3::LearnerRegr | NULL)
Regression learner for the surrogate model filtering algorithm.

Method clone()

The objects of this class are cloneable with this method.

Usage

FiltorSurrogateProgressive$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.