Mutator Choosing Action Probabilistically

Mutator that chooses which operation to perform probabilistically. The Mutator wraps two other Mutators given during construction, and for each individuum, the operation to perform is sampled: with probability p (configuration parameter), the Mutator given to the mutator construction argument is applied, and with probability p - 1 the one given to mutator_not is applied.

Configuration Parameters

This operator has the configuration parameters of the Mutators that it wraps: The configuration parameters of the operator given to the mutator construction argument are prefixed with "maybe.", the configuration parameters of the operator given to the mutator_not construction argument are prefixed with "maybe_not.".

Additional configuration parameters:

p :: numeric(1)
Probability per individual with which to apply the operator given to the mutator construction argument.

Supported Operand Types

Supported Param classes are the set intersection of supported classes of mutator and mutator_not.

Dictionary

This Mutator can be created with the short access form mut() (muts() to get a list), or through the the dictionary dict_mutators in the following way:

# preferred:
mut("maybe", <mutator> [, <mutator_not>])
muts("maybe", <mutator> [, <mutator_not>])  # takes vector IDs, returns list of Mutators

# long form:
dict_mutators$get("maybe", <mutator> [, <mutator_not>])

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Mutator -> MutatorMaybe

Methods

Public methods

MutatorMaybe$new()

MutatorMaybe$prime()

MutatorMaybe$clone()

Inherited methods

miesmuschel::MiesOperator$operate()

Method new()

Initialize the MutatorMaybe object.

Usage

MutatorMaybe$new(mutator, mutator_not = MutatorNull$new())

Arguments

mutator

(Mutator)
Mutator to wrap. This operator gets run with probability p (Configuration parameter).
The constructed object gets a clone of this argument.

mutator_not

(Mutator)
Another Mutator to wrap. This operator runs when mutator is not chosen. By default, this is MutatorNull, i.e. no operation. With this default, the MutatorMaybe object applies the mutator operation with probability p, and no operation at all otherwise.
The constructed object gets a clone of this argument.

Method prime()

See MiesOperator method. Primes both this operator, as well as the wrapped operators given to mutator and mutator_not during construction.

Usage

MutatorMaybe$prime(param_set)

Arguments

param_set

(ParamSet)
Passed to MiesOperator$prime().

Returns

invisible self.

Method clone()

The objects of this class are cloneable with this method.

Usage

MutatorMaybe$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

set.seed(1)
mm = mut("maybe", mut("gauss", sdev = 5), p = 0.5)
p = ps(x = p_int(-5, 5), y = p_dbl(-5, 5))
data = data.frame(x = rep(0, 5), y = rep(0, 5))

mm$prime(p)
mm$operate(data)

#>    x           y
#> 1  5  2.07320717
#> 2 -5 -4.64283517
#> 3  0  0.00000000
#> 4  0  0.00000000
#> 5 -1 -0.02883586

mm$param_set$values$p = 0.3
mm$operate(data)

#>   x         y
#> 1 0  0.000000
#> 2 0  0.000000
#> 3 0  0.000000
#> 4 0  0.000000
#> 5 2 -3.106203

mm2 = mut("maybe",
  mutator = mut("gauss", sdev = 0.01),
  mutator_not = mut("gauss", sdev = 10),
  p = 0.5
)

mm2$prime(p)
mm2$operate(data)

#>    x            y
#> 1  0 -0.008919211
#> 2  0 -0.012375384
#> 3 -1  5.000000000
#> 4  0  0.003773956
#> 5  0  0.008041895