Crossover Recombinator

Values between two individuals are exchanged with component-wise independent probability.

Configuration Parameters

p :: numeric(1)
Component-wise probability with which to exchange values. Initialized to 0.5.

Supported Operand Types

Supported Param classes are: ParamLgl, ParamInt, ParamDbl, ParamFct

Dictionary

This Recombinator can be created with the short access form rec() (recs() to get a list), or through the the dictionary dict_recombinators in the following way:

# preferred:
rec("xounif")
recs("xounif")  # takes vector IDs, returns list of Recombinators

# long form:
dict_recombinators$get("xounif")

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Recombinator -> RecombinatorCrossoverUniform

Methods

Public methods

RecombinatorCrossoverUniform$new()

RecombinatorCrossoverUniform$clone()

Inherited methods

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$prime()

Method new()

Initialize the RecombinatorCrossoverUniform object.

Usage

RecombinatorCrossoverUniform$new(keep_complement = TRUE)

Arguments

keep_complement

(logical(1))
Whether the operation should keep both individuals that were crossed over (TRUE), or only the first and discard the crossover complement (FALSE). Default TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

RecombinatorCrossoverUniform$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

set.seed(1)
rx = rec("xounif")
p = ps(x = p_int(-5, 5), y = p_dbl(-5, 5), z = p_dbl(-5, 5))
data = data.frame(x = 0:5, y = 0:5, z = 0:5)

rx$prime(p)
rx$operate(data)

#>   x y z
#> 1 1 1 0
#> 2 0 0 1
#> 3 2 3 2
#> 4 3 2 3
#> 5 4 4 4
#> 6 5 5 5

rx$param_set$values$p = 0.3
rx$operate(data)

#>   x y z
#> 1 0 0 0
#> 2 1 1 1
#> 3 2 2 3
#> 4 3 3 2
#> 5 4 5 5
#> 6 5 4 4