Multi-Objective Fixed Projection Scalor

Scalor that returns the maximum of a set of projections.

Priming PS must contain a "scalarization_weights" tagged ParamUty that contains weight matrices (Nobjectives x Nweights) or vectors (if Nweights is 1).

Configuration Parameters

scalarization :: function
Function taking a fitness-matrix fitnesses (Nindivs x Nobjectives, with higher values indicating higher desirability) and a list of weight matrices weights (Nindivs elements of Nobjectives x Nweights matrices; positive weights should indicate a positive contribution to scale) and returns a matrix of scalarizations (Nindivs x Nweights, with higher values indicating greater desirability).
While custom functions can be used, it is recommended to use a Scalarizer, such as scalarizer_linear(), or scalarizer_chebyshev().

Supported Operand Types

Supported Param classes are: ParamLgl, ParamInt, ParamDbl, ParamFct

Dictionary

This Scalor can be created with the short access form scl() (scls() to get a list), or through the the dictionary dict_scalors in the following way:

# preferred:
scl("fixedprojection")
scls("fixedprojection")  # takes vector IDs, returns list of Scalors

# long form:
dict_scalors$get("fixedprojection")

See also

Super classes

miesmuschel::MiesOperator -> miesmuschel::Scalor -> ScalorFixedProjection

Active bindings

weights_component_id

(numeric(1))
search space component identifying the weights by which to scalarize.

Methods

Public methods

ScalorFixedProjection$new()

ScalorFixedProjection$prime()

ScalorFixedProjection$clone()

Inherited methods

miesmuschel::MiesOperator$help()

miesmuschel::MiesOperator$operate()

miesmuschel::MiesOperator$print()

miesmuschel::MiesOperator$repr()

Method new()

Initialize the ScalorFixedProjection object.

Usage

ScalorFixedProjection$new(weights_component_id = "scalarization_weights")

Arguments

weights_component_id

(character(1))
Id of the search space component identifying the weights by which to scalarize. Default "scalarization_weights".

Method prime()

See MiesOperator method. Primes both this operator, as well as the operator given to the operation configuration parameter. Note that this modifies the $param_set$values$operation object.

Usage

ScalorFixedProjection$prime(param_set)

Arguments

param_set

(ParamSet)
Passed to MiesOperator$prime().

Returns

invisible self.

Method clone()

The objects of this class are cloneable with this method.

Usage

ScalorFixedProjection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

set.seed(1)