Filter Offspring

Uses a Filtor to extract a subset of individuals from a given set. The individuals are either returned directly (when get_indivs is TRUE) or in form of an index into the given individuals (when get_indivs is FALSE).

Filtors must always select individuals without replacement, so selected individual indices are unique.

mies_filter_offspring(
  inst,
  individuals,
  lambda,
  filtor = NULL,
  fidelity_schedule = NULL,
  budget_id = NULL,
  get_indivs = TRUE
)

Arguments

inst
(OptimInstance)
Optimization instance to evaluate.

individuals
(data.frame | data.table)
Individuals to filter. Must have columns according to filtor$primed_ps, and must have at least filtor$needed_input(lambda) rows.

lambda
(integer(1))
Number of individuals to filter down to.

filtor
(Filtor | NULL)
Filtor operator that filters. When NULL is given, then the FiltorNull operation is performed and the first lambda individuals are taken from individuals.

fidelity_schedule
(data.frame | NULL)
data.frame with three columns "generation", "budget_new", "budget_survivors", in that order. "budget_new" and "budget_survivors" are atomic columns assigned to the budget_id component of offspring; which one is chosen depends on survivor_budget. "generation" is an integer valued column, indicating the first generation at which a row is valid. At least one row with generation == 1 must be present.
This value may be NULL if no multi-fidelity optimization is performed, or when, instead of the current generation's budget, the maximum budget found in inst$archive should be used (the default).
Note that the multifidelity functionality is experimental and the UI may change in the future.

budget_id
(character(1) | NULL)
Budget compnent when doing multi-fidelity optimization. This component of the search space is added to individuals according to fidelity_schedule and the current generation's "budget_survivors". Should be NULL when and only when fidelity_schedule is NULL.

get_indivs
(logical(1))
Whether to return the data.frame or data.table of selected individuals, or an index into individuals.

Value

If get_indivs is TRUE: a data.frame or data.table (depending on the input type of individuals) of filtered configurations. If get_indivs is FALSE: an integer vector indexing the filtered individuals.