Initialize MIES Optimization

Set up an OptimInstance for MIES optimization. This adds the dob and eol columns to the instance's archive, and makes sure there are at least mu survivors (i.e. entries with eol set to NA) present. If there are already >= mu prior evaluations present, then the last mu of these remain alive (the other's eol set to 0); otherwise, up to mu new randomly sampled configurations are evaluated and added to the archive and have eol set to NA.

mies_init_population(
  inst,
  mu,
  initializer = generate_design_random,
  fidelity_schedule = NULL,
  budget_id = NULL,
  additional_component_sampler = NULL
)

Arguments

inst
(OptimInstance)
Optimization instance to evaluate.

mu
(integer(1))
Population target size, non-negative integer.

initializer
(function)
Function that generates a Design object, with arguments param_set and n, functioning like paradox::generate_design_random or paradox::generate_design_lhs. Note that paradox::generate_design_grid can not be used and must be wrapped with a custom function that ensures that only n individuals are produced. The generated design must correspond to the inst's $search_space; for components that are not in the objective's search space, the additional_component_sampler is used.

fidelity_schedule
(data.frame | NULL)
data.frame with three columns "generation", "budget_new", "budget_survivors", in that order. "budget_new" and "budget_survivors" are atomic columns assigned to the budget_id component of offspring; which one is chosen depends on survivor_budget. "generation" is an integer valued column, indicating the first generation at which a row is valid. At least one row with generation == 1 must be present.
This value may be NULL if no multi-fidelity optimization is performed (the default).
Note that the multifidelity functionality is experimental and the UI may change in the future.

budget_id
(character(1) | NULL)
Budget component that is set to the value found in fidelity_schedule when doing multi-fidelity optimization. Should be NULL when and only when fidelity_schedule is NULL.

additional_component_sampler
(Sampler | NULL)
Sampler for components of individuals that are not part of inst's $search_space. These components are never used for performance evaluation, but they may be useful for self-adaptive OperatorCombinations. See the description of mies_prime_operators() on how operators need to be primed to respect additional components.
It is possible that additional_component_sampler is used for more rows than initializer, which happens when the inst's $archive contains prior evaluations that are alive, but does not contain columns pertaining to additional columns, or contains all these columns but there are rows that are NA valued. If only some of the columns are present, or if all these columns are present but there are rows that are only NA valued for some columns, then an error is thrown.
Default is NULL: no additional components.

Value

invisible OptimInstance: the input instance, modified by-reference.

See also

Examples

library("bbotk")
lgr::threshold("warn")

# Define the objective to optimize
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

# Get a new OptimInstance
oi <- OptimInstanceSingleCrit$new(objective,
  terminator = trm("evals", n_evals = 100)
)

mies_init_population(inst = oi, mu = 3)

# 3 evaluations, archive contains 'dob' and 'eol'
oi$archive

#> <Archive>
#>      x    y dob eol  Obj           timestamp batch_nr
#> 1: 2.1 0.99   1  NA 0.71 2021-07-16 05:18:47        1
#> 2: 0.3 2.31   1  NA 0.11 2021-07-16 05:18:47        1
#> 3: 2.6 3.95   1  NA 0.03 2021-07-16 05:18:47        1

###
# Advanced demo, making use of additional components and fidelity_schedule
# Note that the multifidelity functionality is experimental and the UI may
# change in the future.
##

# Get a new OptimInstance
oi <- OptimInstanceSingleCrit$new(objective,
  terminator = trm("evals", n_evals = 100)
)

mies_init_population(inst = oi, mu = 3,
  fidelity_schedule = data.frame(
    generation = c(1, 3),
    budget_new = c(1, 2),
    budget_survivors = c(2, 3)
  ), budget_id = "y",
  additional_component_sampler = Sampler1DRfun$new(
    param = ParamDbl$new("additional", -1, 1), rfun = function(n) rep(-1, n)
  )
)

# 3 evaluations. We also have 'additional', sampled from rfun (always -1),
# which is ignored by the objective. Besides, we have "y", which is 2,
# according to the fidelity_schedule: `budget_survivors` of `generation` 1.
oi$archive

#> <Archive>
#>    additional    x y dob eol  Obj           timestamp batch_nr
#> 1:         -1 0.28 2   1  NA 0.12 2021-07-16 05:18:47        1
#> 2:         -1 2.66 2   1  NA 1.29 2021-07-16 05:18:47        1
#> 3:         -1 3.61 2   1  NA 0.15 2021-07-16 05:18:47        1