Terminator that Counts OptimizerMies Generations

Terminator that terminates after a given number of generations have passed in OptimizerMies.

If OptimizerMies is started on an archive that already has evaluated configurations, these evaluations count as generation 0. If an initial, randomly sampled generation is generated by OptimizerMies, it has generation number 1. Setting generation to 1 therefore terminates after the evaluation of the initial sample, unless no initial sample is generated by OptimizerMies and instead found in the archive. generation set to 0 avoids any evaluation within OptimizerMies (but is ignored if no dob column is in the archive).

When doing multi-fidelity optimization, and fidelity of a configuration is increased because of a step in the fidelity schedule, or because they were sampled new and survived, then this fidelity refinement happens as part of an already started generation. This means termination at this fidelity refinement step is avoided.

Dictionary

This Terminator can be created with the short access form trm() (trms() to get a list), or through the dictionary mlr_terminators in the following way:

# preferred
trm("gens")
trms("gens")  # takes vector IDs, returns list of Terminators

# long form
mlr_terminators$get("gens")

Configuration Parameters

generations :: integer(1)
Number of generations to evaluate, after which to stop. Initialized to Inf.

Super class

bbotk::Terminator -> TerminatorGenerations

Methods

Public methods

TerminatorGenerations$new()

TerminatorGenerations$is_terminated()

TerminatorGenerations$clone()

Inherited methods

bbotk::Terminator$format()

bbotk::Terminator$print()

bbotk::Terminator$remaining_time()

bbotk::Terminator$status()

Method new()

Initialize the TerminatorGenerations object.

Usage

TerminatorGenerations$new()

Method is_terminated()

Is TRUE if when the termination criterion is matched, FALSE otherwise.

Usage

TerminatorGenerations$is_terminated(archive)

Arguments

archive

Archive Archive to check.

Returns

logical(1): Whether to terminate.

Method clone()

The objects of this class are cloneable with this method.

Usage

TerminatorGenerations$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("bbotk")
trm("gens", generations = 10)

#> <TerminatorGenerations>
#> * Parameters: generations=10