Short Access Forms for Operators

These functions complement dict_mutators, dict_recombinators, dict_selectors with functions in the spirit of mlr3::mlr_sugar.

mut(.key, ...)

muts(.keys, ...)

rec(.key, ...)

recs(.key, ...)

sel(.key, ...)

sels(.key, ...)

ftr(.key, ...)

ftrs(.key, ...)

Arguments

.key
(character(1))
Key passed to the respective dictionary to retrieve the object.

...
(named list())
Named arguments passed to the constructor, to be set as parameters in the paradox::ParamSet, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.

.keys
(character())
Keys passed to the respective dictionary to retrieve multiple objects.

Value

Mutator for mut()

list of Mutator for muts()

Recombinator for rec().

list of Recombinator for recs().

Selector for sel().

list of Selector for sels().

Examples

mut("gauss", sdev = 0.5)

#> <MutatorGauss>
#>   Inherits from: <MutatorNumeric>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     endomorphism: active binding
#>     initialize: function () 
#>     is_primed: active binding
#>     operate: function (values, ...) 
#>     param_classes: active binding
#>     param_set: active binding
#>     prime: function (param_set) 
#>     primed_ps: active binding
#>   Private:
#>     .endomorphism: TRUE
#>     .integer_native: FALSE
#>     .mutate: function (values) 
#>     .mutate_numeric: function (values, lowers, uppers) 
#>     .operate: function (values) 
#>     .param_classes: ParamInt ParamDbl
#>     .param_set: ParamSet, R6
#>     .param_set_id: NULL
#>     .param_set_source: NULL
#>     .primed_ps: NULL
#>     deep_clone: function (name, value) 

rec("xounif")

#> <RecombinatorCrossoverUniform>
#>   Inherits from: <Recombinator>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     endomorphism: active binding
#>     initialize: function (keep_complement = TRUE) 
#>     is_primed: active binding
#>     n_indivs_in: active binding
#>     n_indivs_out: active binding
#>     operate: function (values, ...) 
#>     param_classes: active binding
#>     param_set: active binding
#>     prime: function (param_set) 
#>     primed_ps: active binding
#>   Private:
#>     .endomorphism: TRUE
#>     .n_indivs_in: 2
#>     .n_indivs_out: 2
#>     .operate: function (values) 
#>     .param_classes: ParamLgl ParamInt ParamDbl ParamFct
#>     .param_set: ParamSet, R6
#>     .param_set_id: NULL
#>     .param_set_source: NULL
#>     .primed_ps: NULL
#>     .recombine: function (values) 
#>     deep_clone: function (name, value) 

sel("random")

#> <SelectorRandom>
#>   Inherits from: <Selector>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     endomorphism: active binding
#>     initialize: function () 
#>     is_primed: active binding
#>     operate: function (values, ...) 
#>     param_classes: active binding
#>     param_set: active binding
#>     prime: function (param_set) 
#>     primed_ps: active binding
#>     supported: active binding
#>   Private:
#>     .endomorphism: FALSE
#>     .operate: function (values, fitnesses, n_select) 
#>     .param_classes: ParamLgl ParamInt ParamDbl ParamFct
#>     .param_set: ParamSet, R6
#>     .param_set_id: NULL
#>     .param_set_source: NULL
#>     .primed_ps: NULL
#>     .select: function (values, fitnesses, n_select) 
#>     .supported: single-crit multi-crit
#>     deep_clone: function (name, value)