Short Access Forms for Operators

These functions complement dict_mutators, dict_recombinators, dict_selectors with functions in the spirit of mlr3::mlr_sugar.

mut(.key, ...)

muts(.keys, ...)

rec(.key, ...)

recs(.key, ...)

sel(.key, ...)

sels(.key, ...)

scl(.key, ...)

scls(.key, ...)

ftr(.key, ...)

ftrs(.key, ...)

Arguments

.key

(character(1))
Key passed to the respective dictionary to retrieve the object.

...

(named list())
Named arguments passed to the constructor, to be set as parameters in the paradox::ParamSet, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.

.keys

(character())
Keys passed to the respective dictionary to retrieve multiple objects.

Value

Mutator for mut()

list of Mutator for muts()

Recombinator for rec().

list of Recombinator for recs().

Selector for sel().

list of Selector for sels().

Scalor for scl().

list of Scalor for scls().

See also

Examples

mut("gauss", sdev = 0.5)
#> mut("gauss")
#> $param_set:
#>                  id lower upper      levels  value
#>              <char> <num> <num>      <list> <list>
#> 1:             sdev    NA    NA                0.5
#> 2: sdev_is_relative    NA    NA  TRUE,FALSE   TRUE
#> 3: truncated_normal    NA    NA  TRUE,FALSE  FALSE
rec("xounif")
#> rec("cmpmaybe", recombinator = rec("swap"), recombinator_not = rec("null", 
#>     n_indivs_in = 2))
#> $param_set:
#>        id lower upper levels  value
#>    <char> <num> <num> <list> <list>
#> 1:      p    NA    NA           0.5
sel("random")
#> sel("random")
#> $param_set:
#>               id lower upper           levels  value
#>           <char> <num> <num>           <list> <list>
#> 1: sample_unique    NA    NA global,groups,no groups
scl("nondom")
#> scl("nondom")
#> $param_set:
#>              id lower upper                               levels        value
#>          <char> <num> <num>                               <list>       <list>
#> 1:      epsilon    NA    NA                                                 0
#> 2:        nadir    NA    NA                                                  
#> 3:       jitter    NA    NA                           TRUE,FALSE         TRUE
#> 4: scale_output    NA    NA                           TRUE,FALSE         TRUE
#> 5:     tiebreak    NA    NA crowdingdist,hvcontrib,domcount,none crowdingdist